+++ packages/block-editor/src/layouts/flow.js
@@ -18,6 +18,8 @@ export default {
 		style,
 		blockName,
 		hasBlockGapSupport,
+		globalBlockGapValue,
+		viewportOverrides,
 		layoutDefinitions = LAYOUT_DEFINITIONS,
 	} ) {
 		const blockGapStyleValue = getGapCSSValue( style?.spacing?.blockGap );
@@ -31,6 +33,14 @@ export default {
 				blockGapValue = getGapCSSValue( blockGapStyleValue?.top );
 			} else if ( typeof blockGapStyleValue === 'string' ) {
 				blockGapValue = getGapCSSValue( blockGapStyleValue );
+			}
+
+			if (
+				! blockGapValue &&
+				globalBlockGapValue &&
+				viewportOverrides === undefined
+			) {
+				blockGapValue = getGapCSSValue( globalBlockGapValue );
 			}
 		}